+++ src/main/java/com/example/springbootboardjpa/repoistory/UserJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserJpaRepository extends JpaRepository<User, Integer> {
+public interface UserJpaRepository extends JpaRepository<User, Long> {
     Optional<User> findByName(String name);
 }
